requirements.txt
# ===============================
# Core LangChain + LangGraph
# ===============================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ===============================
# LLM Providers
# ===============================

# OpenAI (primary)
langchain-openai>=0.2.0
openai>=1.0.0

# Anthropic (Claude)
langchain-anthropic>=0.1.0
anthropic>=0.20.0

# Google Gemini
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# Mistral
langchain-mistralai>=0.1.0
mistralai>=0.4.0

# ===============================
# Local / Open-source Models
# ===============================
transformers>=4.39.0
accelerate>=0.27.0
torch>=2.1.0

# ===============================
# Embeddings
# ===============================
tiktoken>=0.7.0
sentence-transformers>=2.7.0

# ===============================
# Vector Stores
# ===============================
chromadb>=0.5.0
faiss-cpu>=1.8.0

# ===============================
# Managed Vector Store (Optional)
# ===============================
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0

# ===============================
# Data Ingestion & Parsing
# ===============================
pypdf>=5.0.0

unstructured>=0.15.0
unstructured[pdf]>=0.15.0
unstructured[docx]>=0.15.0
unstructured[html]>=0.15.0

# Image & OCR
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.9.0

# Audio transcription
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

# ===============================
# Config & Validation
# ===============================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ===============================
# Utilities & Performance
# ===============================
numpy>=1.26.0
scipy>=1.12.0
tqdm>=4.66.0
tenacity>=8.2.0
orjson>=3.9.15
diskcache>=5.6.3

# ===============================
# Observability & Debugging
# ===============================
loguru>=0.7.2
rich>=13.7.0

opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0

# ===============================
# API Layer (Optional)
# ===============================
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0

# ===============================
# Database (MongoDB)
# ===============================
pymongo[srv]>=4.7.0     # [srv] enables Atlas mongodb+srv:// URIs

# ===============================
# LangSmith Tracing
# ===============================
langsmith>=0.1.0
